server/requirements.txt
fastapi>=0.115,<1.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
uvicorn[standard]>=0.30,<1.0
asyncpg>=0.29,<1.0
tortoise-orm>=0.21,<1.0
aerich>=0.3,<1.0
